backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
psycopg2-binary==2.9.5
gunicorn==20.1.0